requirements.txt
firebase-admin
pyrebase4
flask